Add glob and remove loadMocks
