+++ examples/fixtures/javascript/javascript.as.js
@@ -1,0 +1,7 @@
+async function plugin (fastify, opts) {
+  fastify.get('/javascript', async (req, reply) => {
+    return { hello: 'Javascript example' }
+  })
+}
+
+export default plugin
