+++ examples/fixtures/nested/nested.as.ts
