+++ packages/core/package.json
@@ -30,6 +30,7 @@
     "fastify": "^4.10.2",
     "fastify-tsconfig": "^1.0.1",
     "fs-extra": "^11.1.0",
+    "glob": "^8.1.0",
     "ts-import": "4.0.0-beta.10"
   },
   "devDependencies": {
@@ -37,6 +38,7 @@
     "@sinclair/typebox": "^0.25.13",
     "@tsconfig/node18-strictest-esm": "^1.0.1",
     "@types/fs-extra": "^11.0.1",
+    "@types/glob": "^8.0.0",
     "@types/node": "^18.11.15",
     "tsup": "^6.5.0",
     "typescript": "^4.9.4"
